requirements.txt
torch==2.6.0
torchvision==0.21.0
wandb==0.19.10
onnx==1.17.0
onnxruntime==1.21.1
onnxsim==0.4.36
opencv-python
PyYAML
scikit-learn
pandas
seaborn
matplotlib
tqdm
